# Device Clusterization Reorganization, main branch (2024.04.17.)