--- a/core/include/traccc/clusterization/impl/sparse_ccl.ipp
+++ b/core/include/traccc/clusterization/impl/sparse_ccl.ipp
@@ -51,7 +51,8 @@ TRACCC_HOST_DEVICE inline bool is_adjacent(const traccc::cell& a,
 
 TRACCC_HOST_DEVICE inline bool is_far_enough(const traccc::cell& a,
                                              const traccc::cell& b) {
-    return (a.channel1 - b.channel1) > 1 || a.module_link != b.module_link;
+
+    return (a.channel1 > (b.channel1 + 1)) || (a.module_link != b.module_link);
 }
 
 TRACCC_HOST_DEVICE inline unsigned int sparse_ccl(