--- a/core/include/traccc/clusterization/impl/sparse_ccl.ipp
+++ b/core/include/traccc/clusterization/impl/sparse_ccl.ipp
@@ -44,8 +44,8 @@ TRACCC_HOST_DEVICE inline unsigned int make_union(
 TRACCC_HOST_DEVICE inline bool is_adjacent(const traccc::cell& a,
                                            const traccc::cell& b) {
 
-    return (a.channel0 - b.channel0) * (a.channel0 - b.channel0) <= 1 and
-           (a.channel1 - b.channel1) * (a.channel1 - b.channel1) <= 1 and
+    return (a.channel0 - b.channel0) * (a.channel0 - b.channel0) <= 1 &&
+           (a.channel1 - b.channel1) * (a.channel1 - b.channel1) <= 1 &&
            a.module_link == b.module_link;
 }