--- a/device/common/include/traccc/clusterization/device/form_spacepoints.hpp
+++ b/device/common/include/traccc/clusterization/device/form_spacepoints.hpp
@@ -1,6 +1,6 @@
 /** TRACCC library, part of the ACTS project (R&D line)
  *
- * (c) 2022 CERN for the benefit of the ACTS project
+ * (c) 2022-2024 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -9,14 +9,9 @@
 
 // Project include(s).
 #include "traccc/definitions/qualifiers.hpp"
-#include "traccc/device/fill_prefix_sum.hpp"
 #include "traccc/edm/cell.hpp"
 #include "traccc/edm/measurement.hpp"
 #include "traccc/edm/spacepoint.hpp"
-
-// Vecmem include(s).
-#include <vecmem/containers/data/vector_view.hpp>
-#include <vecmem/memory/memory_resource.hpp>
 
 // System include(s).
 #include <cstddef>
@@ -34,10 +29,10 @@ namespace traccc::device {
 ///
 TRACCC_HOST_DEVICE
 inline void form_spacepoints(
-    const std::size_t globalIndex,
+    std::size_t globalIndex,
     measurement_collection_types::const_view measurements_view,
     cell_module_collection_types::const_view modules_view,
-    const unsigned int measurement_count,
+    unsigned int measurement_count,
     spacepoint_collection_types::view spacepoints_view);
 
 }  // namespace traccc::device