--- a/device/common/include/traccc/clusterization/device/impl/ccl_kernel.ipp
+++ b/device/common/include/traccc/clusterization/device/impl/ccl_kernel.ipp
@@ -33,11 +33,12 @@ namespace traccc::device {
 /// @param[in] barrier  A generic object for block-wide synchronisation
 ///
 template <typename barrier_t>
-TRACCC_DEVICE void fast_sv_1(index_t* f, index_t* gf,
-                             unsigned char adjc[MAX_CELLS_PER_THREAD],
-                             index_t adjv[MAX_CELLS_PER_THREAD][8],
-                             const index_t tid, const index_t blckDim,
-                             barrier_t& barrier) {
+TRACCC_DEVICE void fast_sv_1(
+    details::index_t* f, details::index_t* gf,
+    unsigned char adjc[details::MAX_CELLS_PER_THREAD],
+    details::index_t adjv[details::MAX_CELLS_PER_THREAD][8],
+    const details::index_t tid, const details::index_t blckDim,
+    barrier_t& barrier) {
     /*
      * The algorithm finishes if an iteration leaves the arrays unchanged.
      * This varible will be set if a change is made, and dictates if another
@@ -59,12 +60,13 @@ TRACCC_DEVICE void fast_sv_1(index_t* f, index_t* gf,
          * cluster ID if it is lower than ours, essentially merging the two
          * together.
          */
-        for (index_t tst = 0; tst < MAX_CELLS_PER_THREAD; ++tst) {
-            const index_t cid = tst * blckDim + tid;
+        for (details::index_t tst = 0; tst < details::MAX_CELLS_PER_THREAD;
+             ++tst) {
+            const details::index_t cid = tst * blckDim + tid;
 
             __builtin_assume(adjc[tst] <= 8);
             for (unsigned char k = 0; k < adjc[tst]; ++k) {
-                index_t q = gf[adjv[tst][k]];
+                details::index_t q = gf[adjv[tst][k]];
 
                 if (gf[cid] > q) {
                     f[f[cid]] = q;
@@ -80,8 +82,9 @@ TRACCC_DEVICE void fast_sv_1(index_t* f, index_t* gf,
         barrier.blockBarrier();
 
 #pragma unroll
-        for (index_t tst = 0; tst < MAX_CELLS_PER_THREAD; ++tst) {
-            const index_t cid = tst * blckDim + tid;
+        for (details::index_t tst = 0; tst < details::MAX_CELLS_PER_THREAD;
+             ++tst) {
+            const details::index_t cid = tst * blckDim + tid;
             /*
              * The second stage is shortcutting, which is an optimisation that
              * allows us to look at any shortcuts in the cluster IDs that we
@@ -98,8 +101,9 @@ TRACCC_DEVICE void fast_sv_1(index_t* f, index_t* gf,
         barrier.blockBarrier();
 
 #pragma unroll
-        for (index_t tst = 0; tst < MAX_CELLS_PER_THREAD; ++tst) {
-            const index_t cid = tst * blckDim + tid;
+        for (details::index_t tst = 0; tst < details::MAX_CELLS_PER_THREAD;
+             ++tst) {
+            const details::index_t cid = tst * blckDim + tid;
             /*
              * Update the array for the next generation, keeping track of any
              * changes we make.
@@ -122,23 +126,25 @@ TRACCC_DEVICE void fast_sv_1(index_t* f, index_t* gf,
 
 template <typename barrier_t>
 TRACCC_DEVICE inline void ccl_kernel(
-    const index_t threadId, const index_t blckDim, const unsigned int blockId,
+    const details::index_t threadId, const details::index_t blckDim,
+    const unsigned int blockId,
     const cell_collection_types::const_view cells_view,
     const cell_module_collection_types::const_view modules_view,
-    const index_t max_cells_per_partition,
-    const index_t target_cells_per_partition, unsigned int& partition_start,
-    unsigned int& partition_end, unsigned int& outi, index_t* f, index_t* gf,
+    const details::index_t max_cells_per_partition,
+    const details::index_t target_cells_per_partition,
+    unsigned int& partition_start, unsigned int& partition_end,
+    unsigned int& outi, details::index_t* f, details::index_t* gf,
     barrier_t& barrier, measurement_collection_types::view measurements_view,
-    unsigned int& measurement_count,
     vecmem::data::vector_view<unsigned int> cell_links) {
 
-    // Get device copy of input parameters
+    // Construct device containers around the views.
     const cell_collection_types::const_device cells_device(cells_view);
     const cell_module_collection_types::const_device modules_device(
         modules_view);
     measurement_collection_types::device measurements_device(measurements_view);
 
-    const unsigned int num_cells = cells_device.size();
+    const cell_collection_types::const_device::size_type num_cells =
+        cells_device.size();
 
     /*
      * First, we determine the exact range of cells that is to be examined
@@ -183,12 +189,13 @@ TRACCC_DEVICE inline void ccl_kernel(
         }
         partition_start = start;
         partition_end = end;
+        assert(partition_start <= partition_end);
     }
 
     barrier.blockBarrier();
 
     // Vector of indices of the adjacent cells
-    index_t adjv[MAX_CELLS_PER_THREAD][8];
+    details::index_t adjv[details::MAX_CELLS_PER_THREAD][8];
     /*
      * The number of adjacent cells for each cell must start at zero, to
      * avoid uninitialized memory. adjv does not need to be zeroed, as
@@ -196,32 +203,33 @@ TRACCC_DEVICE inline void ccl_kernel(
      * is set.
      */
     // Number of adjacent cells
-    unsigned char adjc[MAX_CELLS_PER_THREAD];
+    unsigned char adjc[details::MAX_CELLS_PER_THREAD];
 
     // It seems that sycl runs into undefined behaviour when calling
     // group synchronisation functions when some threads have already run
     // into a return. As such, we cannot use returns in this kernel.
 
     // Get partition for this thread group
-    const index_t size = partition_end - partition_start;
+    const details::index_t size = partition_end - partition_start;
     assert(size <= max_cells_per_partition);
 
 #pragma unroll
-    for (index_t tst = 0; tst < MAX_CELLS_PER_THREAD; ++tst) {
+    for (details::index_t tst = 0; tst < details::MAX_CELLS_PER_THREAD; ++tst) {
         adjc[tst] = 0;
     }
 
-    for (index_t tst = 0, cid; (cid = tst * blckDim + threadId) < size; ++tst) {
+    for (details::index_t tst = 0, cid; (cid = tst * blckDim + threadId) < size;
+         ++tst) {
         /*
          * Look for adjacent cells to the current one.
          */
-        device::reduce_problem_cell(cells_device, cid, partition_start,
-                                    partition_end, adjc[tst], adjv[tst]);
+        reduce_problem_cell(cells_device, cid, partition_start, partition_end,
+                            adjc[tst], adjv[tst]);
     }
 
 #pragma unroll
-    for (index_t tst = 0; tst < MAX_CELLS_PER_THREAD; ++tst) {
-        const index_t cid = tst * blckDim + threadId;
+    for (details::index_t tst = 0; tst < details::MAX_CELLS_PER_THREAD; ++tst) {
+        const details::index_t cid = tst * blckDim + threadId;
         /*
          * At the start, the values of f and gf should be equal to the
          * ID of the cell.
@@ -248,45 +256,17 @@ TRACCC_DEVICE inline void ccl_kernel(
      * Count the number of clusters by checking how many cells have
      * themself assigned as a parent.
      */
-    for (index_t tst = 0, cid; (cid = tst * blckDim + threadId) < size; ++tst) {
-
-        if (f[cid] == cid) {
-            // Increment the summary values in the header object.
-            vecmem::device_atomic_ref<unsigned int,
-                                      vecmem::device_address_space::local>
-                atom(outi);
-            atom.fetch_add(1);
-        }
-    }
-
-    barrier.blockBarrier();
-
-    /*
-     * Add the number of clusters of each thread block to the total
-     * number of clusters. At the same time, a cluster id is retrieved
-     * for the next data processing step.
-     * Note that this might be not the same cluster as has been treated
-     * previously. However, since each thread block spawns a the maximum
-     * amount of threads per block, this has no sever implications.
-     */
-    if (threadId == 0) {
+    {
         vecmem::device_atomic_ref<unsigned int,
-                                  vecmem::device_address_space::global>
-            atom(measurement_count);
-        outi = atom.fetch_add(outi);
-    }
+                                  vecmem::device_address_space::local>
+            atom(outi);
+        for (details::index_t tst = 0, cid;
+             (cid = tst * blckDim + threadId) < size; ++tst) {
 
-    barrier.blockBarrier();
-
-    /*
-     * Get the position to fill the measurements found in this thread group.
-     */
-    const unsigned int groupPos = outi;
-
-    barrier.blockBarrier();
-
-    if (threadId == 0) {
-        outi = 0;
+            if (f[cid] == cid) {
+                atom.fetch_add(1);
+            }
+        }
     }
 
     barrier.blockBarrier();
@@ -294,21 +274,18 @@ TRACCC_DEVICE inline void ccl_kernel(
     const vecmem::data::vector_view<unsigned short> f_view(
         max_cells_per_partition, &f[0]);
 
-    for (index_t tst = 0, cid; (cid = tst * blckDim + threadId) < size; ++tst) {
+    for (details::index_t tst = 0, cid; (cid = tst * blckDim + threadId) < size;
+         ++tst) {
         if (f[cid] == cid) {
-            /*
-             * If we are a cluster owner, atomically claim a position in the
-             * output array which we can write to.
-             */
-            vecmem::device_atomic_ref<unsigned int,
-                                      vecmem::device_address_space::local>
-                atom(outi);
-            const unsigned int id = atom.fetch_add(1);
-
-            device::aggregate_cluster(cells_device, modules_device, f_view,
-                                      partition_start, partition_end, cid,
-                                      measurements_device[groupPos + id],
-                                      cell_links, groupPos + id);
+            // Add a new measurement to the output buffer. Remembering its
+            // position inside of the container.
+            const measurement_collection_types::device::size_type meas_pos =
+                measurements_device.push_back({});
+            // Set up the newly created measurement.
+            aggregate_cluster(cells_device, modules_device, f_view,
+                              partition_start, partition_end, cid,
+                              measurements_device[meas_pos], cell_links,
+                              meas_pos);
         }
     }
 }