--- a/device/cuda/include/traccc/cuda/finding/finding_algorithm.hpp
+++ b/device/cuda/include/traccc/cuda/finding/finding_algorithm.hpp
@@ -1,6 +1,6 @@
 /** TRACCC library, part of the ACTS project (R&D line)
  *
- * (c) 2023 CERN for the benefit of the ACTS project
+ * (c) 2023-2024 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -109,6 +109,8 @@ class finding_algorithm
     vecmem::copy& m_copy;
     /// The CUDA stream to use
     stream& m_stream;
+    /// Warp size of the GPU being used
+    int m_warp_size;
 };
 
 }  // namespace traccc::cuda