… device/common/include/traccc/clusterization/device/impl/form_spacepoints.ipp
… device/cuda/CMakeLists.txt
… device/cuda/include/traccc/cuda/cca/component_connection.hpp
… device/cuda/include/traccc/cuda/clusterization/clusterization_algorithm.hpp
… device/cuda/include/traccc/cuda/clusterization/experimental/clusterization_algorithm.hpp
… device/cuda/include/traccc/cuda/clusterization/spacepoint_formation_algorithm.hpp
… device/cuda/include/traccc/cuda/fitting/fitting_algorithm.hpp
… device/cuda/include/traccc/cuda/seeding/seed_finding.hpp
… device/cuda/include/traccc/cuda/seeding/spacepoint_binning.hpp
… device/cuda/include/traccc/cuda/seeding/track_params_estimation.hpp
… device/cuda/include/traccc/cuda/utils/definitions.hpp
… device/cuda/include/traccc/cuda/utils/stream.hpp
… device/cuda/src/cca/component_connection.cu
… device/cuda/src/clusterization/clusterization_algorithm.cu
… device/cuda/src/clusterization/experimental/clusterization_algorithm_exp.cu
… device/cuda/src/clusterization/spacepoint_formation_algorithm.cu
… device/cuda/src/finding/finding_algorithm.cu
… device/cuda/src/fitting/fitting_algorithm.cu
… device/cuda/src/seeding/experimental/spacepoint_formation.cu
… device/cuda/src/seeding/seed_finding.cu
… device/cuda/src/seeding/spacepoint_binning.cu
… device/cuda/src/seeding/track_params_estimation.cu
… device/cuda/src/utils/barrier.hpp
… device/cuda/src/utils/cuda_error_handling.cpp
… device/cuda/src/utils/cuda_error_handling.hpp
… device/cuda/src/utils/make_prefix_sum_buff.cu
… device/cuda/src/utils/opaque_stream.cpp
… device/cuda/src/utils/opaque_stream.hpp
… device/cuda/src/utils/stream.cpp
… device/cuda/src/utils/utils.cpp