After #543, this is an even more ambitious PR for splitting `traccc::cuda::clusterization_algorithm` and `traccc::sycl::clusterization_algorithm` in two. (Re-)Introducing the `traccc::cuda::spacepoint_formation_algorithm` and `traccc::sycl::spacepoint_formation_algorithm` algorithms.

The biggest change is in `traccc::device::ccl_kernel`. Since now that function receives a resizable measurement buffer, it no longer deals with figuring out the index of the measurement that it should fill. It just uses vecmem's built-in ability for doing this sort of a thing.

I also:
  - Got rid of `traccc::cuda::experimental::clusterization_algorithm` and `traccc::sycl::experimental::clusterization_algorithm`. The updated algorithms now behave in the way that those "experimental" algorithms were meant to.
    * But I kept `traccc::cuda::experimental::spacepoint_formation` and `traccc::sycl::experimental::spacepoint_formation` around for now. Since the Detray usage in those is still something that I'll want to steal from. :wink:
  - Got rid of some minute duplications in the code. Introducing some new code that the CUDA and SYCL implementations would use from the same place.
  - Did some general code cleanup here and there.

With all of this, I simplified the device code in a few places a bit. So I wanted to see a little more in detail what happens to the performance of the code with these updates. Unfortunately the answer is a bit complicated... With the multi-threaded throughput tests, using 4 CPU threads, I see the following:

| | $\mu$ = 40 | $\mu$ = 140 | $\mu$ = 200 | $\mu$ = 300 |
|--|--|--|--|--|
| Old CUDA | 2285.54 | 1454.86 | 920.469 | 527.403 |
| New CUDA | 1685.72 | 1038.05 | 742.887 | 424.643 |
| Old SYCL | 1568.65 | 1084.27 | 746.129 | 431.996 |
| New SYCL | 1587.81 | 778.757 | 523.391 | 296.209 |

It indeed looks far from great. :frowning: However, when I do the same with the single-threaded throughput tests, I get:

| | $\mu$ = 40 | $\mu$ = 140 | $\mu$ = 200 | $\mu$ = 300 |
|--|--|--|--|--|
| Old CUDA | 1395.47 | 674.289 | 471.894 | 300.83 |
| New CUDA | 1407.7 | 674.373 | 478.964 | 300.309 |
| Old SYCL | 963.808 | 456.35 | 331.508 | 227.2 |
| New SYCL | 971.387 | 390.687 | 283.706 | 187.615 |

The CUDA numbers here look all good, but SYCL is still not great. :worried: But I don't fully understand why.

At first I was convinced that it would be the simplifications that I made in `traccc::device::ccl_kernel` that would be responsible for the slowdown. But not really. Even when I went back to the previous logic for how positions in the output container are chosen, the throughput numbers didn't change much.

I'll just give up for today... But let's discuss about this tomorrow...