@krasznaa 2024-04-18
I'm still stumped. :frowning:

According to NSight Systems the new code does not execute kernels or memory copies any more slowly. If anything, it is even faster than the current code. :thinking:
  - With the current code I get this overall summary for running 1000 $\mu$=200 events on 4 CPU threads:

![image](https://github.com/acts-project/traccc/assets/30694331/b7d32215-df37-48b6-a24f-a1fb857e0c20)

  - With the new code, for the same setup, I get:

![image](https://github.com/acts-project/traccc/assets/30694331/32e15ced-6722-49ce-92e7-cba7af31db5d)

Notice how the CCL kernel became faster with the updates. :thinking: Remember that, while I do introduce an additional algorithm with this re-shuffling, the number of kernels being executed remains the same!

So, since I could just not understand how the throughput test would become slower, while the kernels seemingly become faster, I ran the applications through VTune as well. And there is indeed a smoking gun there. :thinking: The current code achives the following CPU occupancy during the throughput test:

![image](https://github.com/acts-project/traccc/assets/30694331/05b8ad37-064c-4104-85c0-c667688b1b58)

While this PR's code does this:

![image](https://github.com/acts-project/traccc/assets/30694331/2141e76c-dd5b-4b8a-b667-abb728bc0d09)

(The big spike at 1 thread is due to the initialization in both cases.)

However, I can't figure out what is causing this. :frowning: So if anybody has any good idea, I'm all ears. Since the "threading analysis" of VTune is just not revealing anything to me at the moment...