@beomki-yeo 2024-04-18
I don't know why you get rid of the later part of clusterization which counts the measurement.  (I also don't know why you removed the sorting part. It will need to be rolled back if you encounter weird results from tracking finding once everything is in the single piece.)
How are you counting the number of measurements in the `spacepoint_formation`? Have you checked if you have redundant spacepoints during the seeding?