@beomki-yeo 2024-04-18
Out of curiosity have you checked if you are getting the same results from cpu and cuda?