@krasznaa 2024-04-19
So... fun! I now bumped into the same issue that I believe is giving us grief with clusterizing the ODD cells on a GPU. It seems that the code meant to order cells during CSV reading, is not doing its job. :frowning:

I was getting afraid that there would be no thing left to debug... :face_exhaling: