@krasznaa 2024-04-19
Never mind, we don't get two birds with just the one stone... :frowning: I just didn't set up the new assertion correctly in the code...