@stephenswat 2024-04-19
It would be possible - although weird - that the same cell has multiple activations in the same input?