@stephenswat 2024-04-19
Aha!

```
traccc/data/odd/muon100GeV-geant4 $ cat event000000000-cells.csv | cut -d ',' -f 1,3,4 | sort | uniq -c

...

2 2161728508332653312,219,0
```

So there are some duplicate cells in the input data!