@krasznaa 2024-04-19
That was one of my guesses. But, the following is now also failing for me. :confused:

```diff
diff --git a/io/src/read_geometry.cpp b/io/src/read_geometry.cpp
index c88e8fe..f52ce52 100644
--- a/io/src/read_geometry.cpp
+++ b/io/src/read_geometry.cpp
@@ -19,6 +19,7 @@
 #include <vecmem/memory/host_memory_resource.hpp>
 
 // System include(s).
+#include <cassert>
 #include <stdexcept>
 
 namespace {
@@ -50,6 +51,7 @@ read_json_geometry(std::string_view filename) {
         barcode_map = std::make_unique<
             std::map<std::uint64_t, detray::geometry::barcode>>();
         for (const auto& surface : detector.surfaces()) {
+            assert(barcode_map->find(surface.source) == barcode_map->end());
             (*barcode_map)[surface.source] = surface.barcode();
         }
     }
```

I.e. the same surface identifier shows up multiple times from Detray. Could it be that we end up merging modules in some weird way?

Though this is probably just a red herring. At least for the clusterization...