@krasznaa 2024-04-19
> In fact, some are repeated up to 12 times:
> 
> ```
>       7 2089670364539207936,226,0
>       7 2089670364539207936,229,0
>       7 2089670364539207936,230,0
>       7 2089670364539207936,231,0
>       7 2089670364539317760,156,0
>       7 2089670501978336256,132,0
>       8 1224979373522735104,328,703
>       8 1224979373522735104,329,704
>       8 1729382531788187392,177,42
>       8 1729382531788189440,12,27
>       8 1729382531788189440,13,27
>       8 1729382531788189440,14,27
>       8 1729382669227169536,6,87
>       8 1729382669227169536,7,87
>       8 2089670364539207936,227,0
>       8 2089670364539207936,228,0
>       8 2089670501977937664,38,0
>       8 2089670501977937664,39,0
>       8 2089670501977937664,43,0
>       8 2089670501977937664,44,0
>       8 2089670501978336256,129,0
>       9 1729382531788187392,178,42
>       9 1729382531788187392,179,42
>      10 2089670501977937664,40,0
>      11 2089670501977937664,42,0
>      12 2089670501977937664,41,0
> ```

Since you have some "shell magic" at the ready as it seems, could you remove the duplicates from the files and send me the updated TGZ? I'll upload it as a new version.