@stephenswat 2024-04-19
Interestingly, the full TML dataset also seems to have some duplicated hits, although the number is _much_ smaller than it is for the ODD files. Could be that this just succeeded by chance, e.g. there was never a duplicated hit with 8 other hits in the neighbourhood, so there was some buffer in the 8-length array to compensate.