@krasznaa 2024-04-19
There must be some issue in the Acts "digitization" code. :thinking: This behaviour sounds like a plausible outcome if some mistake is made there. (Remember, simulation just tells us how much energy was deposited exactly where in the detector volumes. We then need to turn those energy deposits into information that the real hardware would've read out as well. I.e. we need to "digitize" the simulated data, as we call it in ATLAS.)

Pinging @asalzburger for info. :wink: