@stephenswat 2024-04-19
```
$ build/bin/traccc_seq_example_cuda --input-directory=odd/muon100GeV-geant4/ --detector-file=geometries/odd/odd_geometry_detray.json --use-detray-detector --digitization-file=geometries/odd/odd-digi-geometric-config.json --input-events=10

Running Full Tracking Chain Using CUDA

>>> Detector Options <<<
  Detector file       : geometries/odd/odd_geometry_detray.json
  Material file       : 
  Surface rid file    : 
  Use detray::detector: yes
  Digitization file   : geometries/odd/odd-digi-geometric-config.json
>>> Input Data Options <<<
  Input data format             : csv
  Input directory               : odd/muon100GeV-geant4/
  Number of input events        : 10
  Number of input events to skip: 0
>>> Clusterization Options <<<
  Target cells per partition: 1024
>>> Track Seeding Options <<<
  None
>>> Performance Measurement Options <<<
  Run performance checks: no
>>> Accelerator Options <<<
  Compare with CPU results: no

WARNING: No material in detector
WARNING: No entries in volume finder
Detector check: OK
==> Statistics ... 
- read    792140 cells from 112034 modules
- created (cpu)  0 measurements     
- created (cpu)  0 spacepoints     
- created (cuda) 319346 spacepoints     
- created  (cpu) 0 seeds
- created (cuda) 42343 seeds
==>Elapsed times...
           File reading  (cpu)  1464 ms
         Clusterization (cuda)  42 ms
   Spacepoint formation (cuda)  24 ms
                Seeding (cuda)  47 ms
           Track params (cuda)  3 ms
                     Wall time  1587 ms
```

:smile: