@stephenswat 2024-04-19
See `pcadp04:/mnt/ssd1/sswatman/traccc/data/traccc-data-v6.tar.gz`.